requirements.txt
fastapi
uvicorn
pydantic
python-multipart
aerich
tortoise-orm
aiosqlite